requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.58.5
fsspec==2024.6.1
huggingface-hub==0.33.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
narwhals==1.46.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.1
pillow==11.0.0
plotly==6.2.0
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.53.0
typing_extensions==4.12.2
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
zipp==3.23.0
